Update xgl from commit: 1b74e8535d
* Update due to argument IImage* retired from SignalNativeFence()
* Remove Mutex::Init(), RWLock::Init(), and ConditionVariable::Init() usage
* [Navi21] X-Plane: LLPC performance tuning
* Add scope to some settings
* Add AMD switchable graphics layer to switch AMD Vulkan driver between amdvlk and RADV
* [Navi21] Madmax LLPC performance tuning
* Fix memory alignment for memory dedicated allocation
* [Navi21] Talos Principle: LLPC performance tuning
* Update PAL Interface in Vulkan to 640
* Update Khronos Vulkan Headers to 1.2.164
* Remove DebugReportCallback::Message() and DebugUtilsMessenger::Message() since they are unused
* PhysicalDevice::m_memoryUsageTracker::trackerMutex corrputed
* [Navi21] Rise of Tomb Raider-LLPC performance tuning
* Move spirv-headers from XGL to LLPC
* flags cleanup - meaningless const on return types
* Enable NGG compactionless for GFX10.3+
* [Navi21] F12017 LLPC performance tuning

+++ icd/api/barrier_policy.cpp
@@ -197,7 +197,7 @@ static VK_INLINE uint32_t ImageLayoutToCacheMask(VkImageLayout imageLayout)
 {
     uint32_t cacheMask = 0;
 
-    switch (imageLayout)
+    switch (static_cast<uint32_t>(imageLayout))
     {
     // It is likely more correct to use Pal::CoherAllUsages but it seems better to avoid CPU sync and just use
     // a general set of cache maks with the unsupported masks eventually being masked out by the barrier policy.
